Support `upstreamAuthor` field.
When the person who built the package is not the original author, this field names the original author (according to the porter). In this case we mainly display the upstream author but we add the "ported by" field to the metadata naming the porter.

The verified github and twitter links are always for the person who built the package. At some point we will need some design work to make this clearer, perhaps placing these icons next to the "ported by" field when there is an upstream author.