Improve app author vs. ported by distinction.
We now always put the packager (who signed the package) on the left. If there is an upstream author, then we add "ported by" to the packager name and put the upstream author name on the right.

Also in this change: Make sure the detail fields (license, etc.) get pushed all the way right as in the mockups.