Disable "popular" and remove limits on items displayed.
"Popular" currently doesn't make sense as we do not have stats.

Dynamically calculating the length of a row had responsiveness issues and also made it unclear that the category had more items. Now we display everything but let it overflow off to the right.

Dynamically loading more categories as you scroll was no longer useful now that the entire index is downloaded upfront. It also was confusing because there was no visual indicator that more categories had been loaded other than the scroll bar jumping.

Perhaps one day we'll bring back the progressive loading -- when we have enough apps for that.