Update release process to account for recent Sansdtorm changes.
So, our release process for the app market involves taking a Sandstorm bundle, ripping out the Sandstorm front-end, and inserting the App Market code in its place. It turns out the Sandstorm bundle-runner is an excellent way to run arbitrary Meteor apps for the same reasons that it is an excellent way to run Sandstorm itself:
- It takes care of Mongo, including setting up authentication intelligently, configuring oplog tailing (hard!), and recovering from crashes.
- It auto-restarts Node and Mongo when they crash.
- It makes updates trivial ("sandstorm update <tarball>").
- We have a lot of experience wrangling the Sansdtorm bundle runner, compared to alternatives like mup.

However, the process had been working so well that I became complacent, basically relying on `release.sh` to do the right thing without the need to test. Until today, that had always just worked.

Today, though, there were problems:
- A recent change updated the version of Meteor, which in turn requires a newer version of Node. I had been using an old version of the Sandstorm bundle with an old Node as the basis for Market releases, so the result was a server that refused to start. Unfortunately due to my complacency this problem was not caught until hitting production.
- I frantically built a new Sansdtorm bundle based on the latest Sandstorm code and re-ran the release process, but this uncovered a new problem: Sandstorm's meteor-bundle-main.js had recently become much more complicated and had grown a dependency on node-forge, which was not in the App Market's dependency list. Hence, after building against a newer Sandstorm bundle, the server still failed to start, for a completely different reason.

This change fixes the latter problem by introducing our own meteor-bundle-main.js which we use in place of Sandstorm's. It is based on the old meteor-bundle-main.js from the old bundle that I had been using.