Add new apps to the app list and re-order it in a structured way.
The top apps are based on actual popularity on Oasis, up until the point where this stat gets noisy (i.e. changes frequently).

Following that are apps whose original authors directly maintain the Sandstorm packaging (whether or not the app is exclusive).

Following that are apps packaged by the community (but not by the original authors).

The last section is apps ported (or written) by Sandstorm employees.