Adjust ordering a bit more accurately.
Activity stats on the stats page are ordered by daily actives, but paying attention to weekly/monthly makes more sense. Radicale had an unusual number of daily actives yesterday but it's weekly/monthly stats aren't as impressive (yet).

Piwik also had an anomaly yesterday but doesn't really belong on the popular list.

I also changed my mind about putting WordPress in the popular apps list. It's really not all that great an experience on Sandstorm.

HackerSlides would actually be next on the popular list, whereas we currently have it first on the next list, which turns out to look the same in practice.