Stop showing shortDescription while big app card is loading
Close #95

Note that the shortDescription template helper in this case requires the
renderHighlighted template in order to be rendered properly. Big app cards
are never, I believe, shown in search results. Showing the shortDescription
is actually worse than showing nothing while the big app card is loading
because the flash of text that gets suddenly replaced might make users
think that they missed out on something.