Make market work without DDP.
The app index is served from app-index.sandstorm.io, a completely different server (which is in fact a Sandstorm App!).

We only use DDP to fetch reviews. If we simply remove the reviews then we no longer need DDP at all.

This in turn means we can shut down the app marketplace server and instead serve it as a static web page! That'll save some money.